Move crc header files into the src/lib/crc directory, and move the serial header files into src/lib/serial
