+++ src/arduino/jaiabot_runtime/crc16.h
@@ -1,1 +1,1 @@
-../../lib/lora/crc16.h
+../../lib/crc/crc16.h
